pyproject.toml
[project]
name = "shitbox-telemetry"
version = "0.1.0"
description = "Offline-first rally car telemetry system"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "tgreen" }]

dependencies = [
    "gpsd-py3>=0.3.0",
    "smbus2>=0.4.0",
    "adafruit-circuitpython-mcp9808>=3.3.0",
    "paho-mqtt>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "tenacity>=8.0.0",
    "requests>=2.28.0",
    "python-snappy>=0.6.0",
    "protobuf>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
shitbox = "shitbox.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
